+++ src/components/EventsInfoPublic/EventsInfo.jsx
@@ -5,6 +5,7 @@ import { useState, useEffect } from "react";
 import { db } from "../../Firebase/FirebaseConfig";
 import { collection, getDocs, doc, query, getDoc } from "firebase/firestore";
 import data from "../../data.json";
+import AddToCalendarButton from "../AddToCalendarButton/AddToCalendarButton";
 
 function EventsInfo({ communityId }) {
   const [userAttending, setUserAttending] = useState(false);
@@ -142,6 +143,7 @@ function EventsInfo({ communityId }) {
             />
             {userAttending ? "Attending" : "RSVP"}
           </button>
+        <AddToCalendarButton event={event}/>
         </div>
       ))}
     </>
